+++ dCommon/dEnums/dCommonVars.h
@@ -427,26 +427,6 @@ enum class UseItemResponse : uint32_t {
 	MountsNotAllowed
 };
 
-/**
- * Represents the different types of inventories an entity may have
- */
-enum eInventoryType : uint32_t {
-	ITEMS = 0,
-	VAULT_ITEMS,
-	BRICKS,
-	MODELS_IN_BBB,
-	TEMP_ITEMS = 4,
-	MODELS,
-	TEMP_MODELS,
-	BEHAVIORS,
-	PROPERTY_DEEDS,
-	VENDOR_BUYBACK = 11,
-	HIDDEN = 12, //Used for missional items
-	VAULT_MODELS = 14,
-	ITEM_SETS, //internal
-	INVALID // made up, for internal use!!!
-};
-
 enum eRebuildState : uint32_t {
 	REBUILD_OPEN,
 	REBUILD_COMPLETED = 2,
